+++ Tests/JournalMonitorTests.cs
@@ -325,6 +325,14 @@ namespace UnitTests
         }
 
         [TestMethod]
+        public void TestJournalCommsSystemMessage()
+        {
+            string line = @"{ ""timestamp"":""2018-11-15T06:16:23Z"", ""event"":""ReceiveText"", ""From"":"""", ""Message"":""$COMMS_entered:#name=ICZ JH-V c2-7;"", ""Message_Localised"":""Entered Channel: ICZ JH-V c2-7"", ""Channel"":""npc"" }";
+            List<Event> events = JournalMonitor.ParseJournalEntry(line);
+            Assert.IsTrue(events.Count == 0);
+        }
+
+        [TestMethod]
         public void TestJournalPlayerDirectMessage()
         {
             string line = "{ \"timestamp\":\"2017-10-12T19:58:46Z\", \"event\":\"ReceiveText\", \"From\":\"SlowIce\", \"Message\":\"good luck\", \"Channel\":\"player\" }";